auth/src/api/controllers/__test__/signup.test.ts
import { app } from "../../../app";
import request from "supertest";
import jwt from "jsonwebtoken";

it("returns a 201 on successful signup", async () => {
    return request(app)
        .post("/api/users/signup")
        .send({
            email: "test@test.com",
            password: "1234567",
        })
        .expect(201);
});

it("returns a 400 with an invalid email", async () => {
    return request(app)
        .post("/api/users/signup")
        .send({
            email: "test",
            password: "1234567",
        })
        .expect(400);
});

it("returns a 400 with an invalid password", async () => {
    return request(app)
        .post("/api/users/signup")
        .send({
            email: "test@test.com",
            password: "1",
        })
        .expect(400);
});

it("returns a 400 with missing email and password", async () => {
    await request(app)
        .post("/api/users/signup")
        .send({
            email: "test@test.com",
        })
        .expect(400);

    await request(app)
        .post("/api/users/signup")
        .send({
            password: "1234567",
        })
        .expect(400);
});

it("disallows duplicate emails", async () => {
    await request(app)
        .post("/api/users/signup")
        .send({
            email: "test@test.com",
            password: "1234567",
        })
        .expect(201);

    await request(app)
        .post("/api/users/signup")
        .send({
            email: "test@test.com",
            password: "1234567",
        })
        .expect(400);
});

it("sets a Bearer token after successful signup", async () => {
    const response = await request(app)
        .post("/api/users/signup")
        .send({
            email: "test@test.com",
            password: "1234567",
        })
        .expect(201);

    expect(response.body).toHaveProperty("token");

    const { token } = response.body;
    const payload = jwt.verify(token, process.env.JWT_KEY as string);

    expect(payload).toHaveProperty("id");
    expect(payload).toHaveProperty("email", "test@test.com");
});
